[147AI] 任务 #统一积分换算口径并修复补单不一致
变更类型：修复
适用范围：海外版
问题原因：
- ManualCompleteTopUp() Stripe 补单仍走旧的 Money×QuotaPerUnit 路径，与正常 webhook 入账结果不一致
- 后台手动调额默认按 raw quota 输入，与积分展示口径不一致
- 前端积分除数默认值 20000 与后端实际返回 5000 不一致，导致页面加载时积分显示跳动

修复方式：
- model/topup.go ManualCompleteTopUp() Stripe 分支改为复用 CalcQuotaByAmount()，与 Recharge() 路径统一
- model/topup_quota_custom.go CalcQuotaByAmount 返回值改为 int，消除浮点口径
- 新增 web/src/helpers/creditQuota.js，提供 getCreditDivisor/quotaToCredits/creditsToQuota/formatCredits
- EditUserModal 调整额度改为默认按积分输入，支持积分/金额/raw quota 三套联动
- 多语言补充积分相关 i18n key

测试建议：
- 正常 Stripe 充值与管理员补单同一笔订单，验证到账积分一致
- 后台调整额度输入 6000，确认用户余额增加 6000 积分

+++ model/topup_quota_custom.go
@@ -5,10 +5,10 @@ package model
 // 与 quota/creditQuotaRatio 的余额显示完全一致。
 const creditQuotaRatio = 5000
 
-func CalcQuotaByAmount(topUp *TopUp) float64 {
-	return float64(topUp.Amount) * creditQuotaRatio
+func CalcQuotaByAmount(topUp *TopUp) int {
+	return int(topUp.Amount) * creditQuotaRatio
 }
 
-func CreditDivisor() float64 {
+func CreditDivisor() int {
 	return creditQuotaRatio
 }
